Solution to Check if a Number is Prime or Not

+++ Check if a Number is Prime or Not@@6788dd63490637cfc44e1ba2/code.c
@@ -4,7 +4,7 @@ int main() {
     int n,is_prime =1;
     scanf("%d",&n);
     
-    for(int i =2;i<=sqrt(n);i+=1){
+    for(int i =2;i<=(int)sqrt(n);i+=1){
         if(n%i==0){
             is_prime =0;
             break;
